refactor(commands): moves command handlers to application
