+++ GitOut/App.xaml.cs
@@ -8,7 +8,6 @@ using GitOut.Features.Git;
 using GitOut.Features.Git.RepositoryList;
 using GitOut.Features.Git.Stage;
 using GitOut.Features.Git.Storage;
-using GitOut.Features.GlobalCommands;
 using GitOut.Features.Logging;
 using GitOut.Features.Material.Snackbar;
 using GitOut.Features.Navigation;
@@ -42,7 +41,7 @@ namespace GitOut
 
             ILogger<App> logger = host.Services.GetRequiredService<ILogger<App>>();
             RegisterExceptionHandlers(logger);
-            _ = host.Services.GetServices<IGlobalCommandHandler>();
+            _ = host.Services.GetService<Features.Wpf.Commands.Application>();
 
             IHostApplicationLifetime life = host.Services.GetRequiredService<IHostApplicationLifetime>();
             life.ApplicationStarted.Register(LogLifetimeEvent(logger, "Host started"));
@@ -77,11 +76,11 @@ namespace GitOut
             services.AddSingleton<INavigationService, NavigationService>();
             services.AddSingleton<ITitleService, TitleService>();
             services.AddSingleton<IWritableStorage, FileStorage>();
+            services.AddSingleton<Features.Wpf.Commands.Application>();
 
             services.AddSettingsFeature();
             services.AddGitFeature();
             services.AddThemeFeature();
-            services.AddGlobalHandlersFeature();
 
             services.AddOptions();
             services.AddOptions<NavigationRegistrationOptions>().Configure(options =>
